Flesh out the cracking algorithm. Removed "letter alterations," and simply added capitalization to the rootPassword generation function.
There is an issue with reading over the shorter_wordlist.txt file,
as the for loop seems to end abruptly without cause before its sample
space has been exhausted.

The logic in the applySymbolAlterationsTo() method is not as readable
as it could be, but for effeciency sake it checks the string for the
appropriate characters, and then yields them with the corresponding
replacements.

The printStatistics() function allows the user to gauge the effeciency
of the program, and the progress of the cracking.